+++ src/com/addressbook/service/ContactsService.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface ContactsService {
     void createContact(Contact contact);
     void updateContact(Contact contact);
-    Contact getContact(Long id);
-    void deleteContact(Long id);
+    Contact getContact(Integer id);
+    void deleteContact(Integer id);
     List<Contact> getAll();
 }
